+++ scripts/supermerger.py
@@ -18,7 +18,6 @@ from modules.ui import create_output_panel, create_refresh_button
 import scripts.mergers.mergers
 import scripts.mergers.pluslora
 import scripts.mergers.xyplot
-from scripts.utils.GenParams import Components
 from importlib import reload
 reload(scripts.mergers.mergers)
 reload(scripts.mergers.xyplot)
@@ -31,7 +30,57 @@ from scripts.mergers.model_util import filenamecutter
 
 path_root = basedir()
 
+supermergerui = None
+
+merge = None
+mergeandgen = None
+gen = None
+s_reserve = None
+s_reserve1 = None
+gengrid = None
+s_startreserve = None
+rand_merge = None
+
+msettings = None
+esettings1 = None
+genparams = None
+lucks = None
+currentmodel = None
+dfalse = None
+dtrue = None
+id_sets = None
+xysettings = None
+
+submit_result = None
+imagegal = None
+numaframe = None
+
 def on_ui_tabs():
+    global supermergerui
+
+    global merge
+    global mergeandgen
+    global gen
+    global s_reserve
+    global s_reserve1
+    global gengrid
+    global s_startreserve
+    global rand_merge
+
+    global msettings
+    global esettings1
+    global genparams
+    global lucks
+    global currentmodel
+    global dfalse
+    global dtrue
+    global id_sets
+    global xysettings
+
+    global submit_result
+    global imagegal
+    global numaframe
+
     weights_presets=""
     userfilepath = os.path.join(path_root, "scripts","mbwpresets.txt")
     if os.path.isfile(userfilepath):
@@ -101,9 +150,9 @@ def on_ui_tabs():
 
 
                     with gr.Row():
-                        Components.merge = gr.Button(elem_id="model_merger_merge", value="Merge!",variant='primary')
-                        Components.mergeandgen = gr.Button(elem_id="model_merger_merge", value="Merge&Gen",variant='primary')
-                        Components.gen = gr.Button(elem_id="model_merger_merge", value="Gen",variant='primary')
+                        merge = gr.Button(elem_id="model_merger_merge", value="Merge!",variant='primary')
+                        mergeandgen = gr.Button(elem_id="model_merger_merge", value="Merge&Gen",variant='primary')
+                        gen = gr.Button(elem_id="model_merger_merge", value="Gen",variant='primary')
                         stopmerge = gr.Button(elem_id="stopmerge", value="Stop")
 
                     with gr.Accordion("Generation Parameters",open = False):
@@ -132,7 +181,7 @@ def on_ui_tabs():
 
                     with gr.Accordion("Elemental Merge, Adjust",open = False):
                         with gr.Row():
-                            Components.esettings1 = gr.CheckboxGroup(label = "settings",choices=["print change"],type="value",interactive=True)
+                            esettings1 = gr.CheckboxGroup(label = "settings",choices=["print change"],type="value",interactive=True)
                         with gr.Row():
                             deep = gr.Textbox(label="Blocks:Element:Ratio,Blocks:Element:Ratio,...",lines=2,value="")
                         with gr.Row():    
@@ -148,16 +197,16 @@ def on_ui_tabs():
                     zgrid = gr.Textbox(label="Z grid (Disabled if blank)",lines=3,value="",visible =False)
                     esettings = gr.CheckboxGroup(label = "XYZ plot settings",choices=["swap XY","save model","save csv","save anime gif","not save grid","print change"],type="value",interactive=True)
                     with gr.Row():
-                        Components.gengrid = gr.Button(elem_id="model_merger_merge", value="Sequential XY Merge and Generation",variant='primary')
+                        gengrid = gr.Button(elem_id="model_merger_merge", value="Sequential XY Merge and Generation",variant='primary')
                         stopgrid = gr.Button(elem_id="model_merger_merge", value="Stop XY",variant='primary')
-                        Components.s_reserve1 = gr.Button(value="Reserve XY Plot",variant='primary')
+                        s_reserve1 = gr.Button(value="Reserve XY Plot",variant='primary')
                     dtrue =  gr.Checkbox(value = True, visible = False)                
                     dfalse =  gr.Checkbox(value = False,visible = False)     
                     dummy_t =  gr.Textbox(value = "",visible = False)    
                 blockid=["BASE","IN00","IN01","IN02","IN03","IN04","IN05","IN06","IN07","IN08","IN09","IN10","IN11","M00","OUT00","OUT01","OUT02","OUT03","OUT04","OUT05","OUT06","OUT07","OUT08","OUT09","OUT10","OUT11"]
         
                 with gr.Column(scale = 2):
-                    Components.currentmodel = gr.Textbox(label="Current Model",lines=1,value="")  
+                    currentmodel = gr.Textbox(label="Current Model",lines=1,value="")  
                     submit_result = gr.Textbox(label="Message")
                     mgallery, mgeninfo, mhtmlinfo, mhtmllog = create_output_panel("txt2img", opts.outdir_txt2img_samples)
                     with gr.Accordion("Let the Dice roll",open = False,visible=True):    
@@ -177,7 +226,7 @@ def on_ui_tabs():
                             lucklimits_u = gr.Textbox(label="Upper limit for X",value = "1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1")
                         with gr.Row(): 
                             lucklimits_l = gr.Textbox(label="Lower limit for X",value = "0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0")
-                        Components.rand_merge = gr.Button(elem_id="runrandmerge", value="Run Rand",variant='primary')
+                        rand_merge = gr.Button(elem_id="runrandmerge", value="Run Rand",variant='primary')
 
             with gr.Row(visible = False) as row_inputers:
                 inputer = gr.Textbox(label="",lines=1,value="")
@@ -271,9 +320,9 @@ def on_ui_tabs():
 
             with gr.Tab("Reservation"):
                 with gr.Row():
-                    Components.s_reserve = gr.Button(value="Reserve XY Plot",variant='primary')
+                    s_reserve = gr.Button(value="Reserve XY Plot",variant='primary')
                     s_reloadreserve = gr.Button(value="Reloat List",variant='primary')
-                    Components.s_startreserve = gr.Button(value="Start XY plot",variant='primary')
+                    s_startreserve = gr.Button(value="Start XY plot",variant='primary')
                     s_delreserve = gr.Button(value="Delete list(-1 for all)",variant='primary')
                     s_delnum = gr.Number(value=1, label="Delete num : ", interactive=True, visible = True,precision =0)
                 with gr.Row():
@@ -379,11 +428,11 @@ def on_ui_tabs():
         load_history.click(fn=load_historyf,outputs=[history ])
 
         msettings=[weights_a,weights_b,model_a,model_b,model_c,base_alpha,base_beta,mode,calcmode,useblocks,custom_name,save_sets,id_sets,wpresets,deep,tensor,bake_in_vae]
-        Components.imagegal = [mgallery,mgeninfo,mhtmlinfo,mhtmllog]
-        Components.xysettings=[x_type,xgrid,y_type,ygrid,z_type,zgrid,esettings]
+        imagegal = [mgallery,mgeninfo,mhtmlinfo,mhtmllog]
+        xysettings=[x_type,xgrid,y_type,ygrid,z_type,zgrid,esettings]
         genparams=[prompt,neg_prompt,steps,sampler,cfg,seed,width,height,batch_size]
         hiresfix = [genoptions,hrupscaler,hr2ndsteps,denois_str,hr_scale]
-        Components.lucks = [luckmode,lucksets,lucklimits_u,lucklimits_l,luckseed,luckserial,luckcustom,luckround]
+        lucks = [luckmode,lucksets,lucklimits_u,lucklimits_l,luckseed,luckserial,luckcustom,luckround]
 
         setdefault.click(fn = configdealer,
             inputs =[*genparams,*hiresfix[1:],dfalse],
@@ -398,6 +447,54 @@ def on_ui_tabs():
         s_reverse.click(fn = reversparams,
             inputs =mergeid,
             outputs = [submit_result,*msettings[0:8],*msettings[9:13],deep,calcmode,luckseed,tensor]
+        )
+
+        merge.click(
+            fn=smergegen,
+            inputs=[*msettings,esettings1,*genparams,*lucks,currentmodel,dfalse,*GenParamGetter.txt2img_params],
+            outputs=[submit_result,currentmodel]
+        )
+
+        mergeandgen.click(
+            fn=smergegen,
+            inputs=[*msettings,esettings1,*genparams,*lucks,currentmodel,dtrue,*GenParamGetter.txt2img_params],
+            outputs=[submit_result,currentmodel,*imagegal]
+        )
+
+        gen.click(
+            fn=simggen,
+            inputs=[*genparams,*GenParamGetter.txt2img_params,currentmodel,id_sets],
+            outputs=[*imagegal],
+        )
+
+        s_reserve.click(
+            fn=numanager,
+            inputs=[gr.Textbox(value="reserve",visible=False),*xysettings,*msettings,*genparams,*lucks,*GenParamGetter.txt2img_params],
+            outputs=[numaframe]
+        )
+
+        s_reserve1.click(
+            fn=numanager,
+            inputs=[gr.Textbox(value="reserve",visible=False),*xysettings,*msettings,*genparams,*lucks,*GenParamGetter.txt2img_params],
+            outputs=[numaframe]
+        )
+
+        gengrid.click(
+            fn=numanager,
+            inputs=[gr.Textbox(value="normal",visible=False),*xysettings,*msettings,*genparams,*lucks,*GenParamGetter.txt2img_params],
+            outputs=[submit_result,currentmodel,*imagegal],
+        )
+
+        s_startreserve.click(
+            fn=numanager,
+            inputs=[gr.Textbox(value=" ",visible=False),*xysettings,*msettings,*genparams,*lucks,*GenParamGetter.txt2img_params],
+            outputs=[submit_result,currentmodel,*imagegal],
+        )
+
+        rand_merge.click(
+            fn=numanager,
+            inputs=[gr.Textbox(value="random",visible=False),*xysettings,*msettings,*genparams,*lucks,*GenParamGetter.txt2img_params],
+            outputs=[submit_result,currentmodel,*imagegal],
         )
 
         search.click(fn = searchhistory,inputs=[searchwrods,searchmode],outputs=[history])
